add: Born and Wolf and circular aperture models
Set up the types and generic functions and added the Born and Wolf
model and the circular aperture model.

* Circular aperture model includes curvature for modeling aberration, but
this should be done in a different way. Maybe by having an ideal
circular aperture model and a aberration function that is dependent on
the normalized frequency. This should also become deprecated if I add
pupil models where aberrations can be added using Zernicke polynomials.

* Born and Wolf model is well established but it may include some
mistakes. This should be checked.

* Tests should be added.