feat: Full refactor of the package architecture
Instead of calling psf and otf methods on the models with sizes, there
is a type SampledOTF and SampledPSF (only SampledPSF implemented now)
that are used to sample from the models with the generation of
particular arrays for the image analysis... This is a better system
since the model often stays the same for a particular use but the
sampling may be different. It also makes more sense (hopefully) for
a user to use the sampled transfer functions rather than sampling the
model directly.

TODO: Should probably change the API of the models to be called instead
of a function for otf and or psf. This would be the case if it was
possible to call a psf and otf generation for the same model. This would
be probably implemented such that you could call an `otf_model` function
or `ModelOTF` on a `<:ModelPSF` type to transform it into a OTF model
that can generate the connected OTF instead of the PSF.