fix(compat)!: Restrict to julia ^1.9
Reason for this is the missing method `round(Float64, RoundFromZero)`
which is used in the bead estimation. There are several methods to
avoid restricting the version, but since LTS version is as of today
already 1.10, I am not bound to support any versions that are older
with a compromise on utility.